+++ apps/api/src/kick_logs/application/ports/__init__.py
@@ -1,13 +1,17 @@
 from kick_logs.application.ports.channel_repository import ChannelRepository
 from kick_logs.application.ports.message_repository import MessageRepository
+from kick_logs.application.ports.password_hasher import PasswordHasher
 from kick_logs.application.ports.sender_repository import SenderRepository
+from kick_logs.application.ports.token_service import TokenService
 from kick_logs.application.ports.unit_of_work import UnitOfWork
 from kick_logs.application.ports.user_repository import UserRepository
 
 __all__ = [
     "ChannelRepository",
     "MessageRepository",
+    "PasswordHasher",
     "SenderRepository",
+    "TokenService",
     "UnitOfWork",
     "UserRepository",
 ]